Move Utils to separate files in utils package
