+++ app/src/main/java/pl/pwr/zpi/bcycle/mobile/utils/TripUtils.kt
@@ -1,25 +1,19 @@
-package pl.pwr.zpi.bcycle.mobile
+package pl.pwr.zpi.bcycle.mobile.utils
 
-import io.reactivex.Single
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
 import org.gavaghan.geodesy.Ellipsoid
 import org.gavaghan.geodesy.GeodeticCalculator
 import org.gavaghan.geodesy.GlobalPosition
 import org.threeten.bp.Duration
-import org.threeten.bp.ZoneId
-import org.threeten.bp.ZonedDateTime
-import org.threeten.bp.chrono.IsoChronology
-import org.threeten.bp.format.DateTimeFormatter
-import org.threeten.bp.format.DateTimeFormatterBuilder
-import org.threeten.bp.format.ResolverStyle
+import pl.pwr.zpi.bcycle.mobile.M_TO_KM
 import pl.pwr.zpi.bcycle.mobile.models.OngoingTripEvent
 import pl.pwr.zpi.bcycle.mobile.models.OngoingTripEventType
 import pl.pwr.zpi.bcycle.mobile.models.TripPoint
 
 /** Given a list of TripPoints, get the distance between them in km. */
 fun getDistance(tripPoints: List<TripPoint>): Double {
-    return tripPoints.zipWithNext { a, b -> getDistance(a, b) }.sum()
+    return tripPoints.zipWithNext { a, b ->
+        getDistance(a, b)
+    }.sum()
 }
 
 /** Given two TripPoints, get the distance between them in km. */
@@ -31,7 +25,6 @@ fun getDistance(from: TripPoint, to: TripPoint): Double {
     val pos2 = GlobalPosition(to.latitude, to.longitude, to.altitude ?: 0.0)
     return geoCalc.calculateGeodeticCurve(reference, pos2, pos1).ellipsoidalDistance / M_TO_KM
 }
-
 
 /** Given two OngoingTripEvents, get the distance between them in km. */
 fun getDistance(from: OngoingTripEvent, to: OngoingTripEvent): Double {
@@ -52,23 +45,4 @@ fun getTime(tripPoints: List<TripPoint>): Int {
     return d.seconds.toInt()
 }
 
-fun <T> Single<T>.background(): Single<T>
-    = this
-        .subscribeOn(Schedulers.io())
-        .observeOn(AndroidSchedulers.mainThread())
 
-
-private var isoUtcFormatter =
-        DateTimeFormatterBuilder().parseCaseInsensitive()
-            .append(DateTimeFormatter.ISO_LOCAL_DATE_TIME).appendOffset("+HH:MM:ss", "Z")
-            .toFormatter()
-            .withChronology(IsoChronology.INSTANCE)
-
-fun dateToIso(date: ZonedDateTime): String =
-    date.withZoneSameInstant(ZoneId.of("UTC")).format(isoUtcFormatter)
-
-fun dateFromIso(date: String): ZonedDateTime =
-    ZonedDateTime.parse(date, isoUtcFormatter)
-
-fun localDateFromIso(date: String): ZonedDateTime =
-    ZonedDateTime.parse(date, isoUtcFormatter).withZoneSameInstant(ZoneId.systemDefault())
